Whitelist all calls to update_option() for non-existent options in Core by default, and add filter

+++ option-usage-checker.php
@@ -1,8 +1,8 @@
 <?php
 /**
  * Plugin Name: Option Usage Checker
- * Description: Make sure that options are less than 1MB when adding and updating them. This is specifically intended for when using an external object cache, like Memcached, where 1MB is the max bucket size (e.g. this is the case on WordPress.com). This ensures compatibility with Memcached's 1MB limit. Also make sure that all options are explicitly added before they are updated. Define Option_Usage_Checker_OBJECT_CACHE_BUCKET_MAX_SIZE (number of bytes) in wp-config.php to override the default of 1MB. Option usage violations will throw exceptions if WP_DEBUG is turned on by default; otherwise PHP warnings will be issued.
- * Version: 0.1
+ * Description: Make sure that options are less than 1MB when adding and updating them. This is specifically intended for when using an external object cache, like Memcached, where 1MB is the max bucket size (e.g. this is the case on WordPress.com). This ensures compatibility with Memcached's 1MB limit. Also make sure that all options are explicitly added before they are updated. Define Option_Usage_Checker_OBJECT_CACHE_BUCKET_MAX_SIZE (number of bytes) in wp-config.php to override the default of 1MB. Option usage violations will throw exceptions if WP_DEBUG is turned on by default; otherwise PHP warnings will be issued. Specific instances of update_option() for non-existent options can be whitelisted via the option_usage_checker_whitelisted filter; all core instnaces are whitelisted by default.
+ * Version: 0.2
  * Author: X-Team WP
  * Author URI: http://x-team.com/wordpress/
  * License: GPLv2+
@@ -80,10 +80,55 @@ class Option_Usage_Checker_Plugin {
 	 */
 	function filter_pre_update_option( $value, $name ) {
 		$this->check_option_size( $name, $value );
-		if ( ! $this->option_exists( $name ) ) {
+		if ( ! $this->option_exists( $name ) && ! $this->_is_update_option_call_whitelisted() ) {
 			$this->handle_error( "Option '$name' does not exist. You must call add_option() before you can call update_option().'" );
 		}
 		return $value;
+	}
+
+	/**
+	 * Walk up callstack to where update_option() was called, and determine if
+	 * the add_option()-absent usage is whitelisted. Uses in Core are whitelisted
+	 * by default.
+	 *
+	 * @return bool
+	 */
+	protected function _is_update_option_call_whitelisted() {
+		$whitelisted = false;
+
+		if ( version_compare( PHP_VERSION, '5.2.5', '>=' ) ) {
+			$callstack = debug_backtrace( false );
+		} else {
+			$callstack = debug_backtrace();
+		}
+
+		$callee = null;
+		foreach ( $callstack as $call ) {
+			if ( ! empty( $call['function'] ) && 'update_option' === $call['function'] ) {
+				$callee = $call;
+			}
+		}
+
+		if ( $callee ) {
+			$is_core = (
+				0 === strpos( $callee['file'], trailingslashit( ABSPATH ) . 'wp-admin/' )
+				||
+				0 === strpos( $callee['file'], trailingslashit( ABSPATH ) . 'wp-includes/' )
+			);
+			$whitelisted = $is_core;
+		}
+
+		/**
+		 * Selectively whitelist calls of update_option() on non-extant options.
+		 * Calls done in Core are whitelisted by default.
+		 *
+		 * @param bool $whitelisted
+		 * @param array|null $callee
+		 * @param array $callstack
+		 */
+		$whitelisted = apply_filters( 'option_usage_checker_whitelisted', $whitelisted, $callee, $callstack );
+
+		return $whitelisted;
 	}
 
 	/**
